requirements.txt
Flask==2.2.3
Flask-Login==0.5.1
PyMongo==4.2.0
Werkzeug==2.2.2
google-auth-oauthlib==0.5.2
google-auth==2.10.0
requests==2.28.2
Jinja2==3.1.2
python-dotenv==0.21.0
pandas==2.0.1
